fix: common module deps in tics scan
