+++ common/tests/unit/test_architecture.py
