+++ kubernetes/tox.ini
@@ -11,6 +11,9 @@ tests_path = "{tox_root}/tests"
 scripts_path = "{tox_root}/scripts"
 terraform_path = "{tox_root}/terraform"
 all_path = {[vars]src_path} {[vars]tests_path} {[vars]scripts_path}
+common_path = "{tox_root}/../common"
+common_tests_path = "{[vars]common_path}/tests/unit"
+common_src_path = "{[vars]common_path}/common"
 
 [testenv]
 set_env =
@@ -51,8 +54,13 @@ description = Run unit tests
 commands_pre =
     poetry install --only main,charm-libs,unit
 commands =
+    # Substrate unit tests -> substrate coverage report (kubernetes/src)
     # `--forked` workaround to deal with memory leaks in scenario tests
     poetry run pytest --forked --numprocesses 120 --cov=src --cov-report=xml --ignore={[vars]tests_path}/integration/ {posargs}
+    # Common unit tests (kubernetes subset) -> common coverage report.
+    # `--cov=src` is included so the scenario tests keep covering the kubernetes
+    # charm source; TICS merges these src hits with the substrate report above.
+    poetry run pytest --forked --numprocesses 120 --cov={[vars]common_src_path} --cov=src --cov-report=xml:{[vars]common_path}/coverage.xml {[vars]common_tests_path} --ignore={[vars]common_tests_path}/machines {posargs}
 
 [testenv:integration]
 pass_env =
