+++ machines/tox.ini
@@ -10,6 +10,9 @@ src_path = "{tox_root}/src"
 tests_path = "{tox_root}/tests"
 terraform_path = "{tox_root}/terraform"
 all_path = {[vars]src_path} {[vars]tests_path}
+common_path = "{tox_root}/../common"
+common_tests_path = "{[vars]common_path}/tests/unit"
+common_src_path = "{[vars]common_path}/common"
 
 [testenv]
 set_env =
@@ -50,7 +53,12 @@ description = Run unit tests
 commands_pre =
     poetry install --only main,charm-libs,unit
 commands =
+    # Substrate unit tests -> substrate coverage report (machines/src)
     poetry run pytest --numprocesses 120 --cov=src --cov-report=xml --ignore={[vars]tests_path}/integration/ {posargs}
+    # Common unit tests (machines subset) -> common coverage report.
+    # `--cov=src` is included so the scenario tests keep covering the machines charm
+    # source; TICS merges these src hits with the substrate report above.
+    poetry run pytest --numprocesses 120 --cov={[vars]common_src_path} --cov=src --cov-report=xml:{[vars]common_path}/coverage.xml {[vars]common_tests_path} --ignore={[vars]common_tests_path}/kubernetes {posargs}
 
 [testenv:integration]
 description = Run integration tests
